requirements.txt
streamlit==1.38.0
torch==2.4.0
torchvision==0.19.0
transformers==4.44.0
pillow==10.4.0
numpy==2.1.0
scikit-learn==1.5.1
pandas==2.2.2
spacy==3.7.2